Backend/requirements.txt
# Core dependencies
flask>=3.0.0
flask-cors>=4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0

# LLM Clients
openai>=1.0.0
google-generativeai>=0.3.0
groq>=0.4.0

# RAG Pipeline
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Async HTTP
httpx>=0.24.0
aiohttp>=3.8.0

# CORS
python-multipart>=0.0.6

# MCP (Model Context Protocol)
mcp>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
